+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 - Add a guide entry: How to import my transactions into CoinTracking?
 - Updated to Qt 5.15 from Qt 5.12 for Linux, macOS and Windows
+- Revamped account-info view to show account keypath, scriptType etc.
 
 ## 4.28.2 [released 2021-06-03]
 - Fix a conversion rates updater bug